+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [4.1.1](https://github.com/LabShare/services/compare/v4.1.0...v4.1.1) (2020-03-06)
+
+
+### Bug Fixes
+
+* read auth url from services.auth.url config setting ([1c808f7](https://github.com/LabShare/services/commit/1c808f7df56158d3c7d0c9428fe139c1301ebfbb))
+
+# [4.1.0](https://github.com/LabShare/services/compare/v4.0.11...v4.1.0) (2020-03-05)
+
+
+### Features
+
+* added build cli command ([941c082](https://github.com/LabShare/services/commit/941c0826249cf83f0567323ce6a8b822cce660d0))
+
 ## [4.0.11](https://github.com/LabShare/services/compare/v4.0.10...v4.0.11) (2020-03-03)